% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{AbsoluteError}
\alias{AbsoluteError}
\title{Absolute error from fitted regression model}
\usage{
AbsoluteError(X, beta, y, weight = NULL, threshold = NULL)
}
\arguments{
\item{X}{Numeric matrix.}

\item{beta}{Numeric vector of regression coefficients.}

\item{weight}{Optional vector of observation weights.}

\item{threshold}{Threshold to derive classifications.}
}
\value{
Absolute error.
}
\description{
Absolute error from fitted regression model
}
