% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccuracyStdErrorEstimation.R
\name{AccuracyStdErrorEstimation}
\alias{AccuracyStdErrorEstimation}
\title{Component-wise minimum variance semi-supervised regression.}
\usage{
AccuracyStdErrorEstimation(
  basis_labeled,
  basis_unlabeled,
  X_labeled,
  X_unlabeled,
  y,
  samp_prob,
  min_var_weight,
  beta_SL,
  beta_MV,
  beta_DR,
  resids_beta_SL,
  resids_beta_imp,
  resids_beta_dr,
  proj_dr,
  inverse_information,
  num_resamples = 500,
  threshold = 0.5
)
}
\arguments{
\item{basis_labeled}{Basis matrix for labeled data set.}

\item{basis_unlabeled}{Basis matrix for unlabeled data set.}

\item{X_labeled}{Covariate matrix for labeled data set.}

\item{X_unlabeled}{Covariate matrix for unlabeled data set.}

\item{y}{Numeric outcome vector.}

\item{samp_prob}{Numeric vector of weights.}

\item{min_var_weight}{Minimum variance weight for semi-supervised estimate.}

\item{beta_SL}{Supervised regression coefficient vector.}

\item{beta_MV}{MinVar Semi-supervised regression coefficient vector.}

\item{beta_DR}{Density ratio regression coefficient vector.}

\item{resids_beta_SL}{Residuals from the supervised regression model.}

\item{resids_beta_imp}{Residuals from the imputation model.}

\item{resids_beta_dr}{Residuals from the density ratio estimator.}

\item{proj_dr}{Projection from density ratio estimator.}

\item{inverse_information}{Inverse information  matrix.}

\item{num_resamples}{Number of resamples.}

\item{threshold}{Threshold for over misclassification rate.}
}
\value{
Pertrubed estimates.
}
\description{
Component-wise minimum variance semi-supervised regression.
}
