% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlternativeBasis.R
\name{AlternativeBasis}
\alias{AlternativeBasis}
\title{Computes the natural spline basis.}
\usage{
AlternativeBasis(X, S, num_knots, basis_type = "interact")
}
\arguments{
\item{X}{Covariate matrix.}

\item{S}{Vector with stratum id.}

\item{num_knots}{Number of knots.}

\item{basis_type}{Type of basis to use.}
}
\value{
Matrix containing basis.
}
\description{
Computes the natural spline basis.
}
