% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataGeneration.R
\name{ARoneCovMat}
\alias{ARoneCovMat}
\title{Generate AR-1 Covariance Matrix}
\usage{
ARoneCovMat(p, rho)
}
\arguments{
\item{p}{Dimension of matrix.}

\item{rho}{Covariance parameter.}
}
\value{
An AR-1 covariance matrix.
}
\description{
Generate AR-1 Covariance Matrix
}
