% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossValAccuracy.R
\name{CrossValAccuracy}
\alias{CrossValAccuracy}
\title{CV estimates for brier score (MSE) and misclassification rate (OMR).}
\usage{
CrossValAccuracy(
  basis_labeled,
  basis_unlabeled,
  X_labeled,
  X_unlabeled,
  y,
  samp_prob,
  min_var_weight,
  num_folds = 5,
  reps = 10,
  threshold = 0.5,
  lambda0 = NULL
)
}
\arguments{
\item{basis_labeled}{Basis matrix for labeled data set.}

\item{basis_unlabeled}{Basis matrix for unlabeled data set.}

\item{X_labeled}{Covariate matrix for labeled data set.}

\item{X_unlabeled}{Covariate matrix for unlabeled data set.}

\item{y}{Numeric outcome vector.}

\item{samp_prob}{Numeric vector of weights.}

\item{min_var_weight}{Numeric vector of minimum variance weights.}

\item{num_folds}{Scalar indicating number of folds for CV.}

\item{reps}{Scalar indicating number of repitions for CV.}

\item{threshold}{Threshold for overall misclassification rate.}

\item{lambda0}{Initial lambda for imputation model.}
}
\value{
CV semi-supervised and supervised MSE and OMR.
}
\description{
CV estimates for brier score (MSE) and misclassification rate (OMR).
}
