% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataGeneration.R
\name{DataGeneration}
\alias{DataGeneration}
\title{Generate Data for Simulation Studies}
\usage{
DataGeneration(
  n_lab,
  n_unlab,
  p,
  rho,
  signal = c(1, 1, 0.5, 0.5),
  model_specification = "outcome_correct_imputation_correct",
  num_strata = 2
)
}
\arguments{
\item{n_lab}{Size of labeled data.}

\item{n_unlab}{Size of unlabeled data.}

\item{p}{Number of covariates.}

\item{rho}{Covariance parameter for AR-1 covariance structure.}

\item{signal}{Values of nonzero regression parameters.}

\item{model_specification}{Choice of model specification.}

\item{num_strata}{Number of strata for stratified sampling.}
}
\value{
List of relevant data objects.
}
\description{
Generate Data for Simulation Studies
}
