% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DensityRatioRegression.R
\name{DensityRatioRegression}
\alias{DensityRatioRegression}
\title{Density ratio regression_}
\usage{
DensityRatioRegression(
  basis_labeled,
  basis_unlabeled,
  X_labeled,
  X_unlabeled,
  y,
  samp_prob,
  lambda = NULL
)
}
\arguments{
\item{basis_labeled}{Basis matrix for labeled data set_}

\item{basis_unlabeled}{Basis matrix for unlabeled data set_}

\item{X_labeled}{Covariate matrix for labeled data set_}

\item{X_unlabeled}{Covariate matrix for unlabeled data set_}

\item{y}{Numeric outcome vector_}

\item{samp_prob}{Numeric vector of weights_}

\item{lambda}{Penalization parameter for initial ridge estimator_}
}
\value{
Vector containing regression coefficients_
}
\description{
Density ratio regression_
}
