% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{ExpitDerivative}
\alias{ExpitDerivative}
\title{Derivative of the Expit Function}
\usage{
ExpitDerivative(x, na_correction = F)
}
\arguments{
\item{x}{Numeric vector.}

\item{na_correction}{Whether to convert NAs to zero.}
}
\value{
Numeric vector.
}
\description{
Calculates \eqn{exp(x) /  (1 + exp(x))^2}.
}
