% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntrinsicEffAccuracy.R
\name{IntrinsicEfficientEstBS}
\alias{IntrinsicEfficientEstBS}
\title{Intrinsic efficiency BS.}
\usage{
IntrinsicEfficientEstBS(
  basis_labeled,
  basis_unlabeled,
  X_labeled,
  X_unlabeled,
  y,
  samp_prob,
  indx_mom,
  lambda0 = NULL,
  theta_prelim
)
}
\arguments{
\item{basis_labeled}{Basis matrix for labeled data set.}

\item{X_labeled}{Covariate matrix for labeled data set.}

\item{X_unlabeled}{Covariate matrix for unlabeled data set.}

\item{y}{Numeric outcome vector.}

\item{samp_prob}{Numeric vector of weights.}

\item{indx_mom}{Numeric vector of indices.}

\item{lambda0}{Penalization parameter for initial ridge estimator.}

\item{theta_prelim}{Initial estimate.}

\item{bais_unlabeled}{Basis matrix for unlabeled data set.}
}
\value{
Vector containing regression coefficients.
}
\description{
Intrinsic efficiency BS.
}
