% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{MeanSquaredError}
\alias{MeanSquaredError}
\title{Mean squared error from fitted regression model}
\usage{
MeanSquaredError(X, beta, y, weight = NULL)
}
\arguments{
\item{X}{Numeric matrix.}

\item{beta}{Numeric vector of regression coefficients.}

\item{weight}{Optional vector of observation weights.}
}
\value{
Mean squared error.
}
\description{
Mean squared error from fitted regression model
}
