% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NaturalSplineBasis.R
\name{NaturalSplineBasis}
\alias{NaturalSplineBasis}
\title{Computes the natural spline basis.}
\usage{
NaturalSplineBasis(X, S, num_knots)
}
\arguments{
\item{X}{Covariate matrix.}

\item{S}{Vector with stratum id.}

\item{num_knots}{Number of knots.}
}
\value{
Matrix containing natural spline basis.
}
\description{
Computes the natural spline basis.
}
