% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DensityRatioRegression.R
\name{Newton_DR}
\alias{Newton_DR}
\title{Density ratio regression_}
\usage{
Newton_DR(
  phi.t,
  phi.v,
  weights,
  max.iter = 100,
  tol = 1e-05,
  initial = rep(0, ncol(phi.t)),
  lambda0 = NULL
)
}
\arguments{
\item{basis_labeled}{Basis matrix for labeled data set_}

\item{bais_unlabeled}{Basis matrix for unlabeled data set_}

\item{X_labeled}{Covariate matrix for labeled data set_}

\item{X_unlabeled}{Covariate matrix for unlabeled data set_}

\item{y}{Numeric outcome vector_}

\item{samp_prob}{Numeric vector of weights_}

\item{lambda}{Penalization parameter for initial ridge estimator_}
}
\value{
Vector containing regression coefficients_
}
\description{
Density ratio regression_
}
