% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{OneHotEncoding}
\alias{OneHotEncoding}
\title{One Hot Encoding}
\usage{
OneHotEncoding(s)
}
\arguments{
\item{s}{Numeric vector of interest containing category for each observation.}
}
\value{
Numeric vectors with the corresponding one hot encoding.
}
\description{
One Hot Encoding
}
