% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunStratifiedSSL.R
\name{RunStratifiedSSL}
\alias{RunStratifiedSSL}
\title{Runs the full analysis}
\usage{
RunStratifiedSSL(
  X_labeled,
  X_unlabeled,
  S_labeled,
  S_unlabeled,
  y,
  samp_prob,
  n_lab,
  num_knots = 3,
  basis_type = "NS_basis",
  num_folds = 3,
  my_threshold = 0.5,
  reps = 10,
  num_perts = 500
)
}
\arguments{
\item{X_labeled}{matrix of n*p}

\item{X_unlabeled}{matrix of N*p}

\item{S_labeled}{vector of n*1}

\item{S_unlabeled}{matrix of N*1}

\item{y}{Numeric outcome vector}

\item{samp_prob}{vector n*1}

\item{n_lab}{integer}

\item{num_knots}{number of knots for basis expansion (default is 3)}

\item{basis_type}{type of basis to use (default is 'NS_basis')}

\item{num_folds}{number of folds for cross-validations (default is 3)}

\item{my_threshold}{threshold to determine classification (default is 0.5)}

\item{reps}{number of replications for cross-validation (default is 10)}

\item{num_perts}{number of replications for perturbation (default is 500)}
}
\value{
Perturbed estimates and corresponding residuals
}
\description{
Runs the full analysis
}
\examples{
data(simulation_data)
n_lab <- 400
n_unlab <- 20000
p <- 10
rho <- 0.4
num_strata <- 2
X_labeled <- simulation_data$covariates_lab
X_unlabeled <- simulation_data$covariates_unlab
S_unlabeled <- simulation_data$S_unlab
S_labeled <- simulation_data$S_lab
y <- simulation_data$Y_lab
samp_prob <- simulation_data$samp_prob
basis_type <- 'NS_basis'
reps = 2
num_perts = 2
num_knots <- 3
basis_type <- 'NS_basis'
my_threshold <- 0.5
RunStratifiedSSL(X_labeled, X_unlabeled, S_labeled, S_unlabeled, y, samp_prob, n_lab)
}
