% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SemiSupervisedAccuracy.R
\name{SemiSupervisedAccuracy}
\alias{SemiSupervisedAccuracy}
\title{Semi-supervised accuracy estimation.}
\usage{
SemiSupervisedAccuracy(
  basis_labeled,
  basis_unlabeled,
  X_labeled,
  X_unlabaled,
  y,
  beta_SSL,
  beta_imp,
  ind_est,
  weight = NULL,
  threshold = NULL
)
}
\arguments{
\item{basis_labeled}{Basis matrix for labeled data set.}

\item{X_labeled}{Covariate matrix for labeled data set.}

\item{y}{Numeric outcome vector.}

\item{beta_SSL}{Numeric vector of regression coefficients.}

\item{beta_imp}{Numeric vector of regression coefficients for imputation.}

\item{ind_est}{Numeric vector to indicate inds for refitting model.}

\item{weight}{Numeric vector of observation weights.}

\item{threshold}{Threshold for overall misclassification rate.}

\item{bais_unlabeled}{Basis matrix for unlabeled data set.}

\item{X_unlabeled}{Covariate matrix for unlabeled data set.}
}
\value{
Semi-supervised accuracy.
}
\description{
Semi-supervised accuracy estimation.
}
