% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SemiSupervisedApparentAccuracy.R
\name{SemiSupervisedApparentAccuracy}
\alias{SemiSupervisedApparentAccuracy}
\title{Apparent estimates for brier score (MSE) and misclassification rate (OMR).}
\usage{
SemiSupervisedApparentAccuracy(
  basis_labeled,
  basis_unlabeled,
  X_labeled,
  X_unlabeled,
  y,
  beta_SSL,
  beta_imp,
  samp_prob,
  ind_est = NULL,
  resamp_weight = NULL,
  threshold = 0.5
)
}
\arguments{
\item{basis_labeled}{Basis matrix for labeled data set.}

\item{basis_unlabeled}{Basis matrix for unlabeled data set.}

\item{X_labeled}{Covariate matrix for labeled data set.}

\item{X_unlabeled}{Covariate matrix for unlabeled data set.}

\item{y}{Numeric outcome vector.}

\item{beta_SSL}{Numeric vector of regression coefficients.}

\item{beta_imp}{Numeric vector of regression coefficients for imputation.}

\item{samp_prob}{Numeric vector of weights.}

\item{ind_est}{Optional numeric vector to indicate inds for refitting model.}

\item{resamp_weight}{Numeric vector of resampling weights.}

\item{threshold}{Threshold for overall misclassification rate.}
}
\value{
Semi-supervised MSE and OMR.
}
\description{
Apparent estimates for brier score (MSE) and misclassification rate (OMR).
}
