% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SemiSupervisedMinVarRegression.R
\name{SemiSupervisedMinVarRegression}
\alias{SemiSupervisedMinVarRegression}
\title{Component-wise minimum variance semi-supervised regression.}
\usage{
SemiSupervisedMinVarRegression(
  beta_SSL,
  beta_SL,
  beta_SSL_se,
  beta_SL_se,
  resids_imp,
  resids_beta_SL,
  X_labeled,
  X_unlabeled,
  epsilon = NULL
)
}
\arguments{
\item{beta_SSL}{Semi-supervised regression coefficient vector.}

\item{beta_SL}{Supervised regression coefficient vector.}

\item{beta_SSL_se}{Semi-supervised regression standard error.}

\item{beta_SL_se}{Supervised regression standard error.}

\item{resids_imp}{Residuals from the imputation model.}

\item{resids_beta_SL}{Residuals from the supervised regression model.}

\item{X_labeled}{Covariate matrix for labeled data set.}

\item{X_unlabeled}{Covariate matrix for unlabeled data set.}

\item{epsilon}{Small offset to help with numerical stability.}
}
\value{
List with parameter and SE est as well as estimated min_var_weights.
}
\description{
Component-wise minimum variance semi-supervised regression.
}
