% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SemiSupervisedRegression.R
\name{SemiSupervisedRegression}
\alias{SemiSupervisedRegression}
\title{Semi-supervised regression.}
\usage{
SemiSupervisedRegression(
  basis_labeled,
  basis_unlabeled,
  X_labeled,
  X_unlabeled,
  y,
  samp_prob,
  lambda = NULL
)
}
\arguments{
\item{basis_labeled}{Basis matrix for labeled data set.}

\item{basis_unlabeled}{Basis matrix for unlabeled data set.}

\item{X_labeled}{Covariate matrix for labeled data set.}

\item{X_unlabeled}{Covariate matrix for unlabeled data set.}

\item{y}{Numeric outcome vector.}

\item{samp_prob}{Numeric vector of weights.}

\item{lambda}{Penalization parameter for initial ridge estimator.}
}
\value{
Vector containing regression coefficients.
}
\description{
Semi-supervised regression.
}
