% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegressionStdErrorEstimation.R
\name{StdErrorEstimationDR}
\alias{StdErrorEstimationDR}
\title{Standard error estimation based on influence function expansion
for DR method.}
\usage{
StdErrorEstimationDR(X_labeled, X_unlabeled, y, beta_dr, resid, proj_dr)
}
\arguments{
\item{X_labeled}{Covariate matrix for labeled data set.}

\item{X_unlabeled}{Covariate matrix for unlabeled data set.}

\item{y}{Numeric outcome vector.}

\item{beta_dr}{DR regression parameter estimate to compute the score function.}

\item{resid}{Estimated residual.}

\item{proj_dr}{DR projection.}
}
\value{
Standard error estimate for DR method.
}
\description{
Standard error estimation based on influence function expansion
for DR method.
}
