% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataGeneration.R
\name{StratificationVar}
\alias{StratificationVar}
\title{Generate Stratification Variable}
\usage{
StratificationVar(covariates, num_strata = 2)
}
\arguments{
\item{covariates}{Covariate matrix with < 2 columns.}

\item{num_strata}{Number of stratum (2 or 4).}
}
\value{
A stratification variable.
}
\description{
Generate Stratification Variable
}
