% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeightedRidgeRegression.R
\name{WeightedRidgeRegression}
\alias{WeightedRidgeRegression}
\title{Ridge regression.}
\usage{
WeightedRidgeRegression(
  X,
  y,
  weights,
  weights_mom,
  indx_mom,
  lambda0 = 1e-04,
  initial = rep(0, 1 + ncol(x))
)
}
\arguments{
\item{X}{Covariate matrix.}

\item{y}{Numeric outcome vector.}

\item{weights}{Numeric vector of weights.}

\item{weights_mom}{Numeric vector of mom weights.}

\item{lambda0}{Penalization parameter.}

\item{initial}{Numeric vector with an intial value.}

\item{idx_mom}{Numeric vector of indices.}
}
\value{
Vector containing regression coefficients.
}
\description{
Ridge regression.
}
