% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewtonGlmnet.R
\name{NewtonGlmnet}
\alias{NewtonGlmnet}
\title{Solve a (regularized) least square problem
with expit link under moment constraints.}
\usage{
NewtonGlmnet(
  X,
  y,
  weights,
  weights_mom,
  indx_mom,
  lambda0 = 1e-04,
  max_iter = 100,
  tol = 1e-04,
  initial = rep(0, 1 + ncol(X))
)
}
\arguments{
\item{X}{Covariate matrix.}

\item{y}{Numeric outcome vector.}

\item{weights}{Numeric vector of weights.}

\item{weights_mom}{Numeric vector of mom weights.}

\item{lambda0}{Penalization parameter.}

\item{max_iter}{Max number of iterations.}

\item{tol}{Desired tolerance level.}

\item{initial}{Numeric vector with an intial value.}

\item{idx_mom}{Numeric vector of indices.}
}
\value{
Vector containing regression coefficients.
}
\description{
Solve a (regularized) least square problem
with expit link under moment constraints.
}
